pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "local-ai"
version = "0.1.0"
description = "White-Box Local AI Agent — Offline, Model-Agnostic, CLI-First"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "psutil>=6.0",
    "typer>=0.12",
    "rich>=13.7",
    "langchain>=0.3",
    "langchain-ollama>=0.2",
    "langchain-chroma>=0.1",
    "langchain-community>=0.3",
    "chromadb>=0.5",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
]

[project.scripts]
local-ai = "src.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
